Refactoring: Move BM model classes into model package. Adjust visibility.
