+++ core/src/main/java/bisq/core/dao/burningman/BtcFeeReceiverService.java
@@ -17,6 +17,7 @@
 
 package bisq.core.dao.burningman;
 
+import bisq.core.dao.burningman.model.BurningManCandidate;
 import bisq.core.dao.state.DaoStateListener;
 import bisq.core.dao.state.DaoStateService;
 import bisq.core.dao.state.model.blockchain.Block;
