+++ desktop/src/main/java/bisq/desktop/main/dao/burnbsq/burningmen/BurnOutputListItem.java
@@ -19,7 +19,7 @@ package bisq.desktop.main.dao.burnbsq.burningmen;
 
 import bisq.desktop.util.DisplayUtils;
 
-import bisq.core.dao.burningman.BurnOutputModel;
+import bisq.core.dao.burningman.model.BurnOutputModel;
 import bisq.core.util.coin.BsqFormatter;
 
 import java.util.Date;
