+++ desktop/src/main/java/bisq/desktop/main/dao/burnbsq/burningmen/BurningmenView.java
@@ -32,9 +32,9 @@ import bisq.desktop.util.validation.BsqValidator;
 import bisq.core.btc.listeners.BsqBalanceListener;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.dao.DaoFacade;
-import bisq.core.dao.burningman.BurningManCandidate;
 import bisq.core.dao.burningman.BurningManPresentationService;
-import bisq.core.dao.burningman.LegacyBurningMan;
+import bisq.core.dao.burningman.model.BurningManCandidate;
+import bisq.core.dao.burningman.model.LegacyBurningMan;
 import bisq.core.dao.governance.proofofburn.ProofOfBurnService;
 import bisq.core.dao.governance.proposal.TxException;
 import bisq.core.dao.state.DaoStateListener;
