+++ desktop/src/main/java/bisq/desktop/main/dao/burnbsq/burningmen/ReimbursementListItem.java
@@ -19,7 +19,7 @@ package bisq.desktop.main.dao.burnbsq.burningmen;
 
 import bisq.desktop.util.DisplayUtils;
 
-import bisq.core.dao.burningman.ReimbursementModel;
+import bisq.core.dao.burningman.model.ReimbursementModel;
 import bisq.core.util.coin.BsqFormatter;
 
 import java.util.Date;
